GER-952: Delete-project plugin : 500 server error when clicking on delete for a project GER-803: Internal server error returned when pushing a commit previously removed by delete-project-plugin GER-527: Delete Plugin :- Same PopUp text for delete and archive action as for just remove data
This review covers multiple Jiras relating to getting the Delete project plugin working in Gerrit 2.16_WD,

Summary:
500 Server error was being returned as we were not correctly forming the delete HTTP request. A POST was being used
in place of a DELETE. The request building mechanism has been reworked.

The RepositoryRemoveHandler in GitMS was still
dealing with an UNINSTALL_REASON and therefore was not archiving the repository correctly.

ChangeIds in gerrit were not being replicated upon user selecting remove data only option

When the user now selects to clean up repository data and leave replicating in GitMS, they will now get a message
stating that the repository was cleaned up instead of deleted.

Change-Id: I83423ab847d99e19238fcc17e6fa40856b3e1d2b